Time: 21 ms (56.19%), Space: 17.9 MB (68.11%) - LeetHub

+++ 0268-missing-number/0268-missing-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int missingNumber(vector<int>& nums) {
-        int sum1 = nums.size();
+        uint16_t sum1 = nums.size();
         for(int i = 0; i < nums.size(); i++) {
             sum1 += i - nums[i];
         }
